Refactor message segments module into smaller units
